refactor: change styles and moved files
