+++ crm/src/styles/style.css
@@ -1,6 +1,9 @@
+@import url('../components/order/orderInCab/style.css');
+
 * {
   box-sizing: border-box;
 }
+
 html {
   width: 100%;
   height: 100vh;
